Refactor to extract generic functions
